+++ packages/sdk/src/lib/dataProtectorSharing/consumeProtectedData.ts
@@ -17,6 +17,7 @@ import {
   throwIfMissing,
   validateOnStatusUpdateCallback,
   positiveNumberSchema,
+  stringSchema,
 } from '../../utils/validators.js';
 import {
   ConsumeProtectedDataParams,
@@ -26,7 +27,7 @@ import {
   SharingContractConsumer,
 } from '../types/index.js';
 import { IExecConsumer } from '../types/internalTypes.js';
-import { getResultFromCompletedTask } from './getResultFromCompletedTask.js';
+import { getResultFromCompletedTask } from '../dataProtectorCore/getResultFromCompletedTask.js';
 import { getAppWhitelistContract } from './smartContract/getAddOnlyAppWhitelistContract.js';
 import { getSharingContract } from './smartContract/getSharingContract.js';
 import {
@@ -56,6 +57,7 @@ export const consumeProtectedData = async ({
   const vMaxPrice = positiveNumberSchema()
     .label('maxPrice')
     .validateSync(maxPrice);
+  const vPath = stringSchema().label('path').validateSync(path);
   let vApp = addressOrEnsSchema().required().label('app').validateSync(app);
   let vWorkerpool = addressOrEnsSchema()
     .label('workerpool')
@@ -215,7 +217,7 @@ export const consumeProtectedData = async ({
     const { result } = await getResultFromCompletedTask({
       iexec,
       taskId,
-      path,
+      path: vPath,
       pemPrivateKey: privateKey,
       onStatusUpdate: vOnStatusUpdate,
     });